+++ config/backend.php
@@ -3,6 +3,7 @@
 return [
 	'payment_corp_id' => env('BACKEND_PAYMENT_CORP_ID'),
 	'payment_division' => env('BACKEND_PAYMENT_DIVISION',1),
+	'failure_email' => env('BACKEND_FAILURE_EMAIL'),
 	'esi' => [
 		'user_id' => env('BACKEND_ESI_USER_ID',0),
 		'client_id' => env('BACKEND_EVE_ESI_CLIENT_ID'),