+++ resources/views/emails/backend_esi_token_bad.blade.php
@@ -1,0 +1,1 @@
+Backend ESI token is bad!